donwload modified node and pre-build
This has contents of a pending PR to node-core. It effectively allows compilation of all node deps, but no linking the binary and compilation of node's .cc files

The pakcage.json reflects that: it downloads a personal branch, configures node as mentioned above and runs make.

Subsequent action is to run test/smoke/node.js